backend/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
python-logstash==0.4.6